# application/static/swagger.yaml
swagger: "2.0"
info:
  title: "Mechanic Shop API"
  description: "This API is designed for a Mechanic Shop Management System"
  version: "1.0.0"
host: "127.0.0.1:5000" # working on local host
schemes:
- "http"
consumes:
- "application/json"
produces:
- "application/json"

####********************************** Documenting the Customer Route******************************####
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: |
      Enter your bearer token in the format: `Bearer <token>`
  ##========== Create Login for Customer (POST) ==========##
paths:
...
  /service-tickets/with-mechanics:
    post:
      tags:
      - Service Tickets
      summary: Create Service Ticket with Assigned Mechanics
      description: |
        Creates a new service ticket and assigns it to one or more mechanics.
        Requires a valid `customer_id` and a list of `mechanic_ids`.
      parameters:
      - in: body
        name: body
        description: JSON payload to create the service ticket
        required: true
        schema:
          $ref: '#/definitions/CreateServiceTicketMechanicPlayload'
      responses:
        201:
          description: Service ticket created successfully
          schema:
            $ref: '#/definitions/CreateServiceTicketMechanicResponse'
          examples:
            application/json:
              message: "New service ticket details added successfully"
              Service:
                VIN: "TTT567623463432"
                customer:
                  address: "200, circle road"
                  email: "deepa@gmail.com"
                  id: 3
                  name: "Deepa"
                  password: "deepa"
                  phone: "574-212-2675"
                  salary: 11000.0
                customer_issue: "Alternator failure"
                id: 32
                mechanics:
                - address: "1/3, Northeast street"
                  email: "gayuu@gmajil.com"
                  id: 7
                  name: "Gayu"
                  phone: "888-543-6785"
                - address: "101, North street"
                  email: "meena@gmajil.com"
                  id: 5
                  name: "Meenaa"
                  phone: "997-543-6785"
                service_date: "2025-07-20"
        400:
          description: Bad request, validation error, or invalid mechanic/customer ID
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Invalid mechanic id: 123"
              error:
                type: string
                example: Customer id not in an account.

  ##=======Documenting Create Service_ticket without mechanic Route (POST) =======##
  /service-tickets:
    post:
      tags:
      - Service Tickets
      summary: Create Service Ticket (without mechanic)
      description: Creates a new service ticket for a customer without assigning a mechanic. Requires a valid `customer_id`.
      parameters:
      - in: body
        name: body
        description: JSON payload to create the service ticket
        required: true
        schema:
          $ref: '#/definitions/CreateServiceTicketPlayload'
      responses:
        201:
          description: Service ticket created successfully
          schema:
            $ref: '#/definitions/CreateServiceTicketResponse'
          examples:
            application/json:
              message: "New service ticket details added successfully"
              Service:
                VIN: "ds112123343432"
                customer:
                  address: "200, circle road"
                  email: "rosy@gmail.com"
                  id: 4
                  name: "Rosy"
                  password: "rosy"
                  phone: "574-212-2675"
                  salary: 11000.0
                customer_issue: "changing teir"
                customer_id: 4
                id: 35
                mechanics: []
                service_date: "2016-10-20"
        400:
          description: Validation error or customer not found
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Customer id not in an account."

    ##============Documenting Get all Service_tickets Route (GET) ==============##

    get:
      #Get request
      tags:
      - Service Tickets
      summary: "Returns all Service tickets"
      description: "Endpoint to retrieve a list of all Service ticket."
      responses:
        200:
          description: "Retrieved Service_tickets Successfully"
          schema:
            $ref: "#/definitions/AllServiceTicketsResponse"

  ##========== Retrieves all service tickets from single customer (GET) =======##
  /service-tickets/my-tickets:
    get:
      tags:
      - Service Tickets
      summary: Get all service tickets for the authenticated customer
      description: |
        Retrieves all service tickets that belong to the currently authenticated customer.
        Requires a valid JWT access token in the `Authorization` header.
      security:
      - bearerAuth: []
      responses:
        200:
          description: List of service tickets belonging to the customer
          schema:
            type: array
            items:
              $ref: '#/definitions/SingleCustomerAllServicesResponse'

  ##=======Documenting Get Specific Service_ticket Route (GET) ===============##
  /service-tickets/{service_id}:
    get:
      tags:
      - Service Tickets
      summary: "Get a Service ticket by ID"
      description: "Service ticket details are generated based on the Service ID and return a specific Service ticket details."
      parameters:
      - in: path ### passing the path parameter
        name: service_id
        required: true
        schema:
          type: integer
        description: ID of the Service ticket details to retrieve
      responses:
        '200':
          description: Service ticket found
          schema:
            $ref: "#/definitions/SpecificServiceResponse"
          examples:
            application/json:
            - id: 32
              VIN: "TTT567623463432"
              customer_issue: "Alternator failure"
              service_date: "2025-07-20"
              customer:
                id: 3
                name: "Deepa"
                email: "deepa@gmail.com"
                phone: "574-212-2675"
                address: "200, circle road"
                salary: 11000.0
              mechanics:
              - id: 7
                name: "Gayu"
                phone: "888-543-6785"
                address: "1/3, Northeast street"
                email: "gayuu@gmajil.com"

    ##=====Edit a Service Tickets by adding and removing  mechanics details (PUT) =======##

    put:
      tags:
      - Service Tickets
      summary: Edit Service Ticket
      description: |
        Update a service ticket by adding or removing mechanic IDs.
        Pass JSON body with lists: 'add_mechanic_ids' and/or 'remove_mechanic_ids'.
      parameters:
      - in: path
        name: service_id
        required: true
        description: ID of the service ticket to update.
        schema:
          type: integer
      - in: "body"
        name: "body"
        required: true
        description: Service ticket edits, specifying mechanics to add or remove
        schema:
          $ref: '#/definitions/EditServicesMechanicsPayload'
      responses:
        200:
          description: Successfully updated service ticket.
          schema:
            $ref: '#/definitions/EditServicesMechanicsResponse'
          examples:
            application/json:
              Messages: "New service ticket details added successfully and Remove successfully"
              Mechanics:
              - "Successfully added item to the mechanic id: 3"
              - "Successfully added item to the mechanic id: 4"
              - "Succefully removed mechanic id: 1"
              - "Succefully removed mechanic id: 2"
              Service:
                id: 21
                VIN: "TTT567623463432"
                customer_issue: "Alternator failure"
                service_date: "2025-07-20"
                customer:
                  id: 3
                  name: "Deepa"
                  email: "deepa@gmail.com"
                  password: "deepa"
                  phone: "574-212-2675"
                  address: "200, circle road"
                  salary: 11000.0
                mechanics:
                - id: 3
                  name: "Pinky"
                  email: "pinky@gmajil.com"
                  phone: "697-543-6785"
                  address: "2, West street"
                - id: 4
                  name: "Ranirani"
                  email: "rani@gmajil.com"
                  phone: "997-543-6785"
                  address: "100, North street"
        400:
          description: Validation error.
        404:
          description: Service ticket not found.

    ##==============Documenting Delete Service_ticket Route (Delete) ===============##

    delete:
      tags:
      - Service Tickets
      summary: "Enpoint to delete a service ticket by ID"
      description: "Endpoint to delete a service ticket. service details are generated based on the service ID and return a successful deleted service ticket."
      parameters:
      - in: path
        name: service_id
        required: true
        type: integer
        description: ID of the service ticket to delete
      responses:
        200:
          description: "Successfully Deleted Service Ticket"
          schema:
            $ref: "#/definitions/DeleteServiceResponse"
          examples:
            application/json:
              message: "Service Ticket id: 3, Successfully deleted."

  ##===== Documenting Update Specific Service_ticket to assign Mechanic details Using (PUT) ======##
  /service-tickets/{service_id}/assign-mechanic/{mechanic_id}:
    put:
      tags:
      - Service Tickets
      summary: Assign Mechanic to Service Ticket
      description: |
        Assigns a mechanic to an existing service ticket.
        This creates a relationship between the service ticket and the mechanic.
        If the mechanic is already assigned, a 400 error is returned.
      parameters:
      - name: service_id
        in: path
        required: true
        description: ID of the service ticket
        schema:
          type: integer
          example: 12
      - name: mechanic_id
        in: path
        required: true
        description: ID of the mechanic to assign
        schema:
          type: integer
          example: 7
      responses:
        200:
          description: Mechanic successfully assigned to the service ticket
          schema:
            $ref: "#/definitions/ServiceAssignMechanicResponse"
          examples:
            application/json:
              Message: "Successfully added item to the mechanic id is: 7  form service ticket id is: 12."
        400:
          description: Mechanic already assigned or invalid service/mechanic ID
          content:
            application/json:
              examples:
                already_assigned:
                  summary: Mechanic already added
                  value:
                    Message: "Details is already included in this service_tickets."
                invalid_ids:
                  summary: Invalid IDs
                  value:
                    Message: "Invalid service_id or mechanic_id"

  ##========= Documenting Remove Specific Service_ticket and Mechanic Using (PUT) ==============##
  /service-tickets/{service_id}/remove-mechanic/{mechanic_id}:
    put:
      tags:
      - Service Tickets
      summary: Remove Mechanic from Service Ticket
      description: |
        Removes the association between a mechanic and a service ticket.
        Returns an error if the mechanic is not currently assigned to the ticket or if either ID is invalid.
      parameters:
      - name: service_id
        in: path
        required: true
        description: ID of the service ticket
        schema:
          type: integer
          example: 12
      - name: mechanic_id
        in: path
        required: true
        description: ID of the mechanic to remove
        schema:
          type: integer
          example: 5
      responses:
        200:
          description: Mechanic successfully removed from the service ticket
          schema:
            $ref: "#/definitions/RemoveServiceMechanicResponse"
          examples:
            application/json:
              Message: "Succefully removed the mechanic id is: 5 form service ticket id is: 12."
        400:
          description: Invalid input or relationship not found
          content:
            application/json:
              examples:
                invalid_ids:
                  summary: Invalid service or mechanic ID
                  value:
                    Message: "Invalid service_id or mechanic_id"
                not_assigned:
                  summary: Mechanic not assigned to service ticket
                  value:
                    error: "Mechanic is not assigned to this service ticket."

  ##======= Documenting a Specific Service_ticket to add on Inventory details Using (POST) ========##
  /service-tickets/{service_id}/add_part:
    post:
      tags:
      - Service Tickets
      summary: Add inventory part to a service ticket
      description: |
        Adds a single part (inventory item) to an existing service ticket by creating a relationship between the service ticket and the inventory part.
        You must provide both `inventory_id` and a positive `quantity` in the request body. Adds a relationship between a service ticket and the inventory
      parameters:
      - in: path
        name: service_id
        required: true
        description: ID of the service ticket to add the inventory part to.
        schema:
          type: integer
      - in: body
        name: body
        required: true
        description: JSON body containing inventory_id and quantity.
        schema:
          $ref: '#/definitions/ServiceAddPartPayload'
      responses:
        200:
          description: Part added to service ticket successfully.
          schema:
            $ref: '#/definitions/ServiceAddPartResponse'
          examples:
            application/json:
              message: "Part added to service ticket successfully."
              inventory_id: 3
              quantity: 15
              service_ticket_id: 3
        400:
          description: Invalid request or missing/incorrect parameters.
          content:
            application/json:
              schema:
                $ref: '#/definitions/ErrorResponse'

  ####********************************* Documenting the Create Inventory Route*****************************####
  ##=======Documenting Create Inventory Route (POST) =======##
...
definitions:
...
  CreateServiceTicketMechanicPlayload:
    type: object
    properties:
      customer_id:
        type: integer
      mechanic_ids:
        type: array
        items:
          type: integer
        example: [ 1, 2, 3 ]
      service_date:
        type: string
        format: date
        example: "2025-07-11"
      VIN:
        type: string
      customer_issue:
        type: string

    required:
    - VIN
    - service_date
    - customer_issue
    - customer_id
    - mechanic_ids
  CreateServiceTicketMechanicResponse:
    type: object
    properties:
      Message:
        type: string
      id:
        type: integer
      VIN:
        type: string
      service_date:
        type: string
        format: date
        example: "2025-07-11"
      customer_issue:
        type: string
      mechanics:
        type: array
        items:
          $ref: '#/definitions/CreateServiceMechanic'
      customer:
        $ref: '#/definitions/CreateServiceMechanicCustomer'
  CreateServiceMechanic:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      address:
        type: string
      email:
        type: string
      phone:
        type: string
  CreateServiceMechanicCustomer:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      address:
        type: "string"
      phone:
        type: "string"
      salary:
        type: number
        format: float

  ## Create Service Ticket without Mechanic Request and Response
  CreateServiceTicketPlayload:
    type: object
    properties:
      customer_id:
        type: integer
      service_date:
        type: string
        format: date
        example: "2025-07-11"
      VIN:
        type: string
      customer_issue:
        type: string

    required:
    - VIN
    - service_date
    - customer_issue
    - customer_id
  CreateServiceTicketResponse:
    type: object
    properties:
      Message:
        type: string
      id:
        type: integer
      VIN:
        type: string
      service_date:
        type: string
        format: date
        example: "2025-07-11"
      customer_issue:
        type: string
      mechanics:
        type: array
        items:
          $ref: '#/definitions/Mechanic'
        description: List of assigned mechanics. May be empty if none are assigned.
        example: []
      customer:
        $ref: '#/definitions/CreateServiceCustomer'
  CreateServiceCustomer:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      address:
        type: "string"
      phone:
        type: "string"
      salary:
        type: number
        format: float

  ##  ALL Service Tickets Response
  AllServiceTicketsResponse:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: integer
        VIN:
          type: string
        service_date:
          type: string
          format: date
          example: "2025-07-11"
        customer_issue:
          type: string
        mechanics:
          type: array
          items:
            $ref: '#/definitions/AllMechanic'
        customer:
          $ref: '#/definitions/AllCustomer'
  AllMechanic:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      address:
        type: string
      email:
        type: string
      phone:
        type: string
  AllCustomer:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      address:
        type: "string"
      phone:
        type: "string"
      salary:
        type: number
        format: float

  ## All service tickets from single customer Response
  SingleCustomerAllServicesResponse:
    type: object
    properties:
      id:
        type: integer
      VIN:
        type: string
      customer_issue:
        type: string
      service_date:
        type: string
        format: date
        example: "2025-07-20"
      customer:
        $ref: '#/definitions/CustomerAllServices'
      mechanics:
        type: array
        items:
          $ref: '#/definitions/Mechanics'
  CustomerAllServices:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      email:
        type: string
      phone:
        type: string
      address:
        type: string
      salary:
        type: number
        format: float
      password:
        type: string
  Mechanics:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      address:
        type: string
      email:
        type: string

  ## Get Specific Service_ticket Response
  SpecificServiceResponse:
    type: object
    properties:
      id:
        type: integer
      VIN:
        type: string
      customer_issue:
        type: string
      service_date:
        type: string
        format: date
        example: "2025-07-20"
      customer:
        $ref: '#/definitions/SingleServiceCustomer'
      mechanics:
        type: array
        items:
          $ref: '#/definitions/SingleServiceMechanic'
...
  EditServicesMechanicsPayload:
    type: object
    properties:
      add_mechanic_ids:
        type: array
        items:
          type: integer
        description: List of mechanic IDs to add.
      remove_mechanic_ids:
        type: array
        items:
          type: integer
        description: List of mechanic IDs to remove.
    example:
      add_mechanic_ids: [ 1, 2 ]
      remove_mechanic_ids: [ 3 ]

    required:
    - add_mechanic_ids
    - remove_mechanic_ids
  EditServicesMechanicsResponse:
    type: object
    properties:
      Messages:
        type: string
      Mechanics:
        type: array
        items:
          type: string
        description: Detailed messages about what was added or removed.
      Service:
        $ref: '#/definitions/EditService'
...
  DeleteServiceResponse:
    type: "object"
    properties:
      message:
        type: "string"

  ## Service Assign Mechanic Response
  ServiceAssignMechanicResponse:
    type: object
    properties:
      Message:
        type: string
        example: "Successfully added item to the mechanic id is: 7  form service ticket id is: 12."

  ## Remove Service mechanic Response
  RemoveServiceMechanicResponse:
    type: object
    properties:
      Message:
        type: string
        example: "Succefully removed the mechanic id is: 5 form service ticket id is: 12."

  ## Add inventory part to a Existing service ticket Request and Responses
  ServiceAddPartPayload:
    type: object
    properties:
      inventory_id:
        type: integer
        description: The ID of the inventory part to add.
        example: 5
      quantity:
        type: integer
        description: Number of items to add (must be greater than 0).
        example: 2

    required:
    - inventory_id
    - quantity
  ServiceAddPartResponse:
    type: object
    properties:
      Message:
        type: string
        example: Part added to service ticket successfully.
      service_ticket_id:
        type: integer
        example: 10
      inventory_id:
        type: integer
        example: 5
      quantity:
        type: integer
        example: 2
      id:
        type: integer
  ErrorResponse:
    type: object
    properties:
      Message:
        type: string
        example: Both inventory_id and quantity are required.

  ####**********************************Definitions for Inventory Routes******************************####

  ## Create Inventory Request and Responses
...